Cleanup

+++ TomeOfHolding/Controllers/CharacterSheetController.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using TomeOfHolding.BLL;
-using TomeOfHolding.BLL.Exceptions;
 using TomeOfHolding.Models;
 
 namespace TomeOfHolding.Controllers {
@@ -16,42 +15,41 @@ namespace TomeOfHolding.Controllers {
 		// Do we want to have a "Get All" here? Maybe for the GM only? IDK how we would implement that
 		[HttpGet("{id}")]
 		public async Task<IActionResult> GetCharacterSheet(int id) {
-			try {
-				CharacterSheet? characterSheet = await _characterSheetService.GetCharacterSheet(id);
-				return Ok(characterSheet);
-			} catch (NotFoundException) {
-				return NotFound("No character Sheet found with that Character");
+			// Will need to figure out how to process the NotFound response properly
+			CharacterSheet? characterSheet = await _characterSheetService.GetCharacterSheet(id);
+			if (characterSheet == null) {
+				return NotFound("No character sheet found.");
 			}
-
+			return Ok(characterSheet);
 		}
 
 		[HttpPost]
 		public async Task<IActionResult> CreateCharacterSheet(CharacterSheet sheet) {
-			if (!ModelState.IsValid) return BadRequest("Invalid Character Sheet");
 			await _characterSheetService.CreateCharacterSheet(sheet);
 			return CreatedAtAction(nameof(GetCharacterSheet), new { id = sheet.CharacterId }, sheet);
 		}
 
 		[HttpDelete("{id}")]
 		public async Task<IActionResult> DeleteCharacterSheet(int id) {
-			try {
-				await _characterSheetService.DeleteCharacterSheet(id);
-				return NoContent();
-			} catch (NotFoundException e) {
-				return NotFound(e.Message);
+			CharacterSheet? characterSheet = await _characterSheetService.GetCharacterSheet(id);
+			if (characterSheet == null) {
+				return NotFound("No character sheet found.");
 			}
-
+			await _characterSheetService.DeleteCharacterSheet(id);
+			return Ok("CharacterSheet deleted successfully.");
 		}
 
 		[HttpPut("{id}")]
 		public async Task<IActionResult> UpdateCharacterSheet(int id, CharacterSheet sheet) {
-			try {
-				if (id != sheet.CharacterId) return BadRequest("CharacterID mismatch.");
-				await _characterSheetService.UpdateCharacterSheet(sheet);
-				return Ok("CharacterSheet updated");
-			} catch (NotFoundException e) {
-				return NotFound(e.Message);
+			if (id != sheet.CharacterId) {
+				return BadRequest("CharacterSheet ID mismatch.");
 			}
+			CharacterSheet? existingCharacterSheet = await _characterSheetService.GetCharacterSheet(id);
+			if (existingCharacterSheet == null) {
+				return NotFound("No character sheet found.");
+			}
+			await _characterSheetService.UpdateCharacterSheet(sheet);
+			return Ok("CharacterSheet updated");
 		}
 	}
 }
